Split core into its own lib. Add tests using catch.hpp. New buddy allocator.
Catch has a fairly high compile-time overhead, i might switch back to gtest
if i can't share a precompiled header between tests.
All symbols are hidden by default when compiling. See CoreExport.h for more info.
Added a new allocator, based on some blog post. It is not optimised at all but
looks like it works, based on the few tests I wrote for it.
There is still a lot of work to be done to streamline test creation, this will
improve over time, once we settle on a test library and the project is correctly
spit into multiple binaries.