Finally settled for doctest for unit testing.
Ported all the tests to the new architecture.

NOTES:
doctest compiles faster than catch for the same features!
Tests are now regrouped in the same binary for each library of the project to amortize the framework's compile time.
Putting tests directly in the libraries was abandoned because it was too much of a hassle and not really worth it.
Note that Catch and doctest don't support static libraries at the time of this commit.