vulkan: support multiple instances during culling
One dispatch is responsible for culling one instance and appending one
command in the command buffer.
One group inside of a dispatch is doing a cluster of 256 triangles.
All indices are tightly packed in the index buffer created by culling.
There's still an issue with disappearing triangles (a few on the total
mesh).

Add new per instance mesh animation to better see culling results in the
capture.

Add shared lib/ folder for shaders. The 'base.hlsl' header is now
mandatory for every shaders.