vulkan: sample shadow map during rendering.
Add memory barrier for the texture.
Add a sampler for the shadow map texture. It does NOT use SampleCmp for
the moment.
We use a very simple constant depth bias for the shadow, this works for
now but is very naive.
Move the light further away to get a better shadow.