renderer: use meshoptimizer to build alternate index buffer
Note that while this makes indices build more coherent polys,
the vertex cache locality is still trash - do not expect a perf boost
yet.
As a workaround for texturing issues, we also fill empty uv/normal
buffers with garbage if they were not found in the model. This allows
the forward rendering pass to fetch the UVs and Normals at the same
offsets as positions. This is dirty of course!