mesh: rewrite vertex buffer for better fetch perf
Take the meshlet input for meshopt and rewrite a more cache-friendly
vertex buffer out of the output data.
We also compact the index buffer.
Since we don't have a coarse per-meshlet gpu culling pass yet, the
indices are translated to be relative to the mesh start. In the future
we should keep the index buffer small (u8 indices) and grab the index
offset in the meshlet instance data.