vulkan: refactor texture upload code
Separate CPU->Staging copy with Staging->GPU.
Allow GPU copy to record at the start of the frame cmdbuffer.
Re-add microprofile code in render passes.
The profile log is always the global one, this will have to be changed
later when multiple cmdbuffer will be recorded at the same time.