vulkan: use VK_KHR_dynamic_rendering
Port main pass to the new api.
This gets rid of needing to create a VkRenderPass object, we just need
to feed a small additional structure at pipeline creation to specify
attachment formats.
Framebuffers also could be removed, since this is all forwarded to
command buffer record time.
Manual barriers have to be placed now before the render pass, but
that's not much additional work.
Fixed a bunch of barriers that were using the legacy enum values (which
are still valid).