vulkan: port scene depth to frame graph
Added abstraction for 'view aspect'. This is set to a reasonable default
value for now, but it lets you override it in the view for specific
use-cases, like having view on the stencil part of a depth+stencil
texture, etc.
As a bonus, this info lets us build the VkImageSubresourceRange
trivially.