framegraph: generate texture transitions automatically
This is only working for textures and it does a few assumptions and/or
hacks to make it work.
For example, we rely on the declaration of passes to always read the
latest version of a resource, and that resources are recreated every
frame.
This is bound to move a lot, so let's keep that as is for now.
The API is still terrible, but the backend starts to go somewhere.