framegraph: support buffer barriers
Massive changeset was need for that, and some hacks made along the way
aren't fixed yet.

The tracking between buffer/texture is done in the builder usage and
stored at the 'FrameGraph::Resource' level with 'is_texture'.
Some unions were added here and there to support differences between the
two.
The queue index is not moved out of the access struct. It doesn't matter
anyway for now, and doesn't make sense to specify at build-time
directly.
Hopefully we'll let the framegraph do smart things here in the future.
Barrier support is as the same level as textures, with the added hack
that I'm also adding a dummy 'undefined' access to transition out from.
We shouldn't need that, but it makes the code simpler. To revisit later.

The GPUBufferView is notably trash right now, definitely fix that soon.

Some more general cleanup was made there too to remove unused cruft:
* Renamed FrameGraphBuilder to just Builder.
* Stopped using local strings for debug framegraph names.
* Renamed 'buffer' to 'handle' in BufferInfo struct. Matches ImageInfo.
* Lifetime tracking is not done anymore at graph build time. Let's do that
later.