vulkan: use meshlet data to drive culling
Instead of dividing the whole input mesh into packs of 256 triangles
manually, use meshlet data provided by meshoptimizer.

This adds another level of indirection in vertex/index lookup which
makes it hard to read and debug, which sucks but will allow the
following future improvements:
* Meshlet coarse culling
* 8bit index buffers (maybe VK_EXT_index_type_uint8)
* ...and more!

Resized the main culling shader for 64-triangle batches. Meshlets reuse
this size to generate the input data.

Removed some of the earlier WIP for meshlet culling, as it made it more
difficult to grep for relevant code. Restoring this part should be the
next step.

NOTE: meshoptimizer doesn't output anything with vertex reuse in the
current state. This definitely should be investigated soon!