vulkan: implement meshlet coarse culling
Another massive changeset, adding a first pass at culling time that
culls meshlets according on their bounding volumes and the camera
frustum.
There's a hack in there to simplify the culling and serve as a POC.
This pass builds a list of meshlets to be rendered in a unique buffer
per pass.

Then we do a very short dispatch to build the arguments for the triangle
culling pass. Nothing special is done here.

The rest of the culling continues as before, with the exception that
draw command compaction is not needed anymore. We can just simply omit
adding a command when the index count is zero.
A lot of code gets thrown away for this change.