vulkan: Add all-in-one GPU macro for debug marker and profile
This places a CPU+GPU profile scope as well as a debug marker in the
command buffer, perfect for render pass tracking.
Also moved the flip even at the start of the frame to have a better
frame split indicator.
Default colors were swapped, and simpler overrides added for the most
usual use cases.