renderer: support node parenting in the scene graph
This is basic and slow but good enough for now.
Each time we want a tranform matrix we recompute the full matrix chain
and build it as we go.
As a proof of concept I added the player mesh as a child node of the
physics node of the scene