vulkan: allow icosahedron as light mesh proxy
Add a vertex buffer for this in the lighting pass.
Add a radius field for point lights. Right now this will result in a
hard cut falloff during lighting. Will need to revisit later.
Support multiple lights during light raster.
Add a light in the test scene as well for this.
Completely fake normals for the sake of simplicity for now.