vulkan: improve swapchain format selection code
No longer requires a preferred format to work, we can choose a decent
one on our own.
HDR is priotized for now. Because reasons.
There's less confusion with the format of the swapchain and the view on
the swapchain, they are both in the same struct and are processed at
similar locations in the code.
ImGUI now shows you which format is selected.